Worked on customizing amplify auth UI.
It isn't as customizable as we need it to be, nor does it offer all the features we need. And it has some bugs. We'll use if for now as we develop, but will need to create custom components for sign in/up, forgot password, confirm email, and socail sign ins.